submit and create destructiln list via react form
